homework_02. add java based config
