+++ lesson1/src/main/java/ru/otus/config/AppConfig.java
@@ -1,0 +1,36 @@
+package ru.otus.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
+
+import java.util.Locale;
+
+@Configuration
+@PropertySource(value = "classpath:application.properties", encoding = "UTF-8")
+public class AppConfig {
+
+
+    @Bean
+    public  static PropertySourcesPlaceholderConfigurer propertySourceConfig() {
+        return new PropertySourcesPlaceholderConfigurer();
+    }
+
+    @Bean
+    public MessageSource messageSource() {
+        ReloadableResourceBundleMessageSource ms = new ReloadableResourceBundleMessageSource();
+        ms.setBasename("bundle");
+        ms.setDefaultEncoding("UTF-8");
+        return ms;
+    }
+
+    @Bean
+    public Locale locale(@Value("${locale.prop}") String locale) {
+        return new Locale(locale);
+    }
+
+}
